fix(manual-trade): keep leverage cap integral

+++ src/client/quantara_app/lib/features/auto_trade/domain/manual_trade_execution.dart
@@ -410,10 +410,9 @@ abstract final class ManualTradeSizingPolicy {
       );
     }
 
-    final maximumPermittedLeverage = math.min(
-      setup.maximumSafeLeverage,
-      rules.maximumLeverage,
-    );
+    final maximumPermittedLeverage = math
+        .min(setup.maximumSafeLeverage, rules.maximumLeverage)
+        .toInt();
     if (maximumPermittedLeverage < rules.minimumLeverage ||
         requestedLeverage < rules.minimumLeverage ||
         requestedLeverage > maximumPermittedLeverage) {
